# feat: alternative optimized proto -> messagepack path for traces