## Which problem is this PR solving?
Unpacking large OTLP payloads is hella slow and allocates gobs of memory. We need to avoid handling vast `map[string]any` full of values we probably won't use.

## Short description of the changes
The main thing here is a giant slab of partially machine-generated code which essentially combines the work done by `proto.Unmarshal` and `TranslateTraceRequest` into a single step, while storing the attribute map in serialized messagepack instead of the more traditional `map[string]any`. Why messagepack? Because it's reasonably compact while still very cheap to read (unlike protobuf), and refinery now has a highly optimized path for handling messagepack data through its entire pipeline. Ultimately, so will shepherd.

This also fixes some pretty weird behavior with nested maps where the 6th level deep would not be consistently json-encoded as you'd expect. Porting the existing broken behavior to the new code was more than even my black heart could take.

Lest this seem like a bonkers thing to do, here is the different it makes for refinery's event ingestion when switched to this code:
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkRouterBatch/otlp     187927        26483         -85.91%

benchmark                     old MB/s     new MB/s     speedup
BenchmarkRouterBatch/otlp     73.91        524.44       7.10x

benchmark                     old allocs     new allocs     delta
BenchmarkRouterBatch/otlp     1790           90             -94.97%

benchmark                     old bytes     new bytes     delta
BenchmarkRouterBatch/otlp     254493        21399         -91.59%
```